[#38] refactor: 컨벤션 수정사항 적용/layout.tsx에 modal-root추가
