+++ src/components/FloatingBox/ParticipantSelector.tsx
