+++ src/components/FloatingBox/PriceDisplay.tsx
@@ -1,6 +1,6 @@
 export default function PriceDisplay() {
   return (
-    <div className='mb-6 mt-15'>
+    <div className='mt-15 mb-6'>
       <div className='mb-1 text-2xl font-bold text-black'>
         ₩ 1,000 <span className='text-xl font-bold text-gray-800'>/ 인</span>
       </div>
