+++ src/components/FloatingBox/TabletPopup.tsx
@@ -1,4 +1,4 @@
-import Calendar from './Calendar';
+import DatePicker from '../DatePicker/DatePicker';
 
 export default function TabletPopup() {
   return (
@@ -8,7 +8,7 @@ export default function TabletPopup() {
           'relative m-auto flex h-fit max-h-[85%] w-screen min-w-[375] flex-col bg-white p-8 px-10 shadow-2xl inset-shadow-sm inset-shadow-gray-300 md:h-fit md:w-[50%] md:max-w-600'
         }
       >
-        <Calendar />
+        <DatePicker />
       </div>
     </div>
   );
