+++ src/components/FloatingBox/TotalPriceDisplay.tsx
