--- src/components/floating-box/styles/DatePickerStyles.tsx
@@ -1,10 +1,0 @@
-export const DatepickerStyles = () => (
-  <style jsx global>{`
-    /* 날짜 하이라이트용 */
-    .highlight-day {
-      border-radius: 9999px;
-      background-color: #3b82f6; /* bg-blue-500 */
-      color: #ffffff;
-    }
-  `}</style>
-);
