+++ src/hooks/useDeviceSize.ts
@@ -1,28 +1,29 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 
 type DeviceType = 'mobile' | 'tablet' | 'desktop';
 
 export default function useDeviceSize(): DeviceType {
-  const getDeviceType = () => {
-    const width = window.innerWidth;
-    if (width < 768) return 'mobile';
-    else if (width < 1024) return 'tablet';
-    else return 'desktop';
-  };
-
-  const [device, setDevice] = useState<DeviceType>(getDeviceType());
+  const [device, setDevice] = useState<DeviceType>('desktop');
 
   useEffect(() => {
-    const checkDevice = () => {
+    const getDeviceType = () => {
       const width = window.innerWidth;
-      if (width < 768) setDevice('mobile');
-      else if (width < 1024) setDevice('tablet');
-      else setDevice('desktop');
+      if (width < 768) return 'mobile';
+      else if (width < 1024) return 'tablet';
+      else return 'desktop';
     };
 
-    checkDevice();
-    window.addEventListener('resize', checkDevice);
-    return () => window.removeEventListener('resize', checkDevice);
+    if (typeof window !== 'undefined') {
+      const handleResize = () => {
+        setDevice(getDeviceType());
+      };
+
+      handleResize(); // 초기 설정
+      window.addEventListener('resize', handleResize);
+      return () => window.removeEventListener('resize', handleResize);
+    }
   }, []);
 
   return device;
