+++ src/ui/BookingModal.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import useDeviceType from '@/hooks/useDeviceSize';
 import MobileModal from './MobileBookingModal';
 import TabletModal from './TabletBookingModal';
 import useDeviceSize from '@/hooks/useDeviceSize';
